Add locally scoped react-select declaration to ndla select
