+++ packages/select/src/Select/types.ts
@@ -6,19 +6,6 @@
  *
  */
 
-import { GroupBase } from 'react-select';
-
-declare module 'react-select/dist/declarations/src/Select' {
-  export interface Props<Option, IsMulti extends boolean, Group extends GroupBase<Option>> {
-    small?: boolean;
-    colorTheme: Color;
-    outline?: boolean;
-    prefix?: string;
-    postfix?: string;
-    bold?: boolean;
-  }
-}
-
 export type SingleValue = Option | null;
 export type MultiValue = readonly Option[];
